+++ rollup.config.js
@@ -1,15 +1,15 @@
 import { ScriptTarget } from 'typescript';
 import path from 'path';
 import multiInput from 'rollup-plugin-multi-input';
-import postcss, { addCssImports } from '@alfalab/rollup-plugin-postcss';
+import postcss, { addCssImports, generateClassNameHash } from '@alfalab/rollup-plugin-postcss';
 import typescript from '@wessberg/rollup-plugin-ts';
-import stringHash from 'string-hash';
 import copy from 'rollup-plugin-copy';
 import json from '@rollup/plugin-json';
 
 import {
     coreComponentsRootPackageResolver,
     coreComponentsResolver,
+    coreComponentsThemesResolver,
 } from './tools/rollup/core-components-resolver';
 import ignoreCss from './tools/rollup/ignore-css';
 import processCss from './tools/rollup/process-css';
@@ -44,19 +44,17 @@ const copyPlugin = dest =>
 const postcssPlugin = postcss({
     modules: {
         generateScopedName: function(name, fileName) {
-            const folderName = path.basename(path.dirname(fileName));
+            const relativeFileName = path.relative(currentPackageDir, fileName);
 
-            const str = `${pkg.name}@${pkg.version}@${folderName}`;
-
-            const hash = stringHash(str)
-                .toString(36)
-                .substr(0, 5);
+            const hash = generateClassNameHash(pkg.name, pkg.version, relativeFileName);
 
             return `${currentComponentName}__${name}_${hash}`;
         },
     },
     extract: true,
     separateCssFiles: true,
+    packageName: pkg.name,
+    packageVersion: pkg.version,
 });
 
 /**
@@ -177,6 +175,9 @@ const esm = {
 
 const rootDir = `../../dist/${currentComponentName}`;
 
+/**
+ * Сборка рут-пакета
+ */
 const root = {
     input: ['dist/**/*.js'],
     external: baseConfig.external,
@@ -205,4 +206,36 @@ const root = {
     ],
 };
 
-export default [es5, modern, cssm, esm, root];
+const configs = [es5, modern, cssm, esm, root];
+
+const themes = pkg.buildThemes;
+
+if (themes && themes.length > 0) {
+    const output = themes.map(theme => ({
+        dir: `dist/themes/${theme}`,
+    }));
+
+    const dest = themes.map(theme => `dist/themes/${theme}`);
+
+    /**
+     * Копирование собранных пакетов в отдельные папки с темами.
+     * В этих папках css-переменные будут заменены на соответствующие значения (после сборки).
+     */
+    configs.push({
+        input: ['dist/**/*.js'], // TODO: убрать cssm
+        external: baseConfig.external,
+        plugins: [
+            multiInput({
+                relative: 'dist',
+            }),
+            copy({
+                flatten: false,
+                targets: [{ src: ['dist/**/*', '!**/*.js'], dest }],
+            }),
+            coreComponentsThemesResolver(),
+        ],
+        output,
+    });
+}
+
+export default configs;