+++ packages/attach/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.16](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@1.4.15...@alfalab/core-components-attach@1.4.16) (2021-03-19)
+
+**Note:** Version bump only for package @alfalab/core-components-attach
+
+
+
+
+
+## [1.4.15](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@1.4.13...@alfalab/core-components-attach@1.4.15) (2021-03-18)
+
+
+### Bug Fixes
+
+* one more sborka bug ([#579](https://github.com/alfa-laboratory/core-components/issues/579)) ([9fbe0be](https://github.com/alfa-laboratory/core-components/commit/9fbe0beca56ec5971de78b3f6cda25305b260efc))
+
+
+
+
+
 ## [1.4.13](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-attach@1.4.11...@alfalab/core-components-attach@1.4.13) (2021-03-16)