+++ packages/button/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-button@2.1.4...@alfalab/core-components-button@2.1.5) (2021-03-19)
+
+
+### Bug Fixes
+
+* **button:** fixed alignment of several buttons in a row ([#561](https://github.com/alfa-laboratory/core-components/issues/561)) ([8d9e1e2](https://github.com/alfa-laboratory/core-components/commit/8d9e1e2f7a4ba8c5c986bb833f7424b38601d463))
+
+
+
+
+
+## [2.1.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-button@2.1.2...@alfalab/core-components-button@2.1.4) (2021-03-18)
+
+
+### Bug Fixes
+
+* one more sborka bug ([#579](https://github.com/alfa-laboratory/core-components/issues/579)) ([9fbe0be](https://github.com/alfa-laboratory/core-components/commit/9fbe0beca56ec5971de78b3f6cda25305b260efc))
+
+
+
+
+
 ## [2.1.2](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-button@2.1.0...@alfalab/core-components-button@2.1.2) (2021-03-16)